start do client

+++ requests/initial.go
@@ -3,36 +3,36 @@ package requests
 import "encoding/binary"
 
 type Initial struct {
-	HeaderSize uint32
-	ReqType    uint16
-	DataSize   uint64
-	NameSize   uint16
+	HeaderSize int32
+	ReqType    int16
+	DataSize   int64
+	NameSize   int16
 	Name       string
 }
 
 // NewInitial
 // err FileNameTooLargeError if len(name) > MaxFileNameSize
 // err FileSizeTooLargeError if len(name) > MaxFileSize
-func NewInitial(dataSize uint64, name string) (req *Initial, err error) {
+func NewInitial(dataSize int64, name string) (req *Initial, err error) {
 	nameSize := len(name)
-	if uint64(nameSize) > uint64(MaxFileNameSize) {
+	if int64(nameSize) > int64(MaxFileNameSize) {
 		return nil, FileNameTooLargeError
 	}
-	if dataSize > uint64(MaxFileSize) {
+	if dataSize > MaxFileSize {
 		return nil, FileSizeTooLargeError
 	}
 
 	return &Initial{
-		HeaderSize: InitialSize(uint16(nameSize)),
+		HeaderSize: InitialSize(int16(nameSize)),
 		ReqType:    InitialReq,
 		DataSize:   dataSize,
-		NameSize:   uint16(nameSize),
+		NameSize:   int16(nameSize),
 		Name:       name,
 	}, nil
 }
 
-func InitialSize(nameSize uint16) uint32 {
-	return uint32(4 + 2 + 8 + 2 + 2 + nameSize)
+func InitialSize(nameSize int16) int32 {
+	return int32(4 + 2 + 8 + 2 + 2 + nameSize)
 }
 
 func (r *Initial) CodeTo(data []byte) (err error) {
@@ -42,10 +42,10 @@ func (r *Initial) CodeTo(data []byte) (err error) {
 		return BufferTooSmallError
 	}
 	dc := binary.BigEndian
-	dc.PutUint32(data[0:4], r.HeaderSize)
-	dc.PutUint16(data[4:6], r.ReqType)
-	dc.PutUint64(data[6:14], r.DataSize)
-	dc.PutUint16(data[14:16], r.NameSize)
+	dc.PutUint32(data[0:4], uint32(r.HeaderSize))
+	dc.PutUint16(data[4:6], uint16(r.ReqType))
+	dc.PutUint64(data[6:14], uint64(r.DataSize))
+	dc.PutUint16(data[14:16], uint16(r.NameSize))
 	copy(data[16:size], r.Name)
 	return nil
 }
@@ -56,10 +56,10 @@ func (r *Initial) DecodeFrom(data []byte) (err error) {
 		return BufferTooSmallError
 	}
 	dc := binary.BigEndian
-	r.HeaderSize = dc.Uint32(data[0:4])
-	r.ReqType = dc.Uint16(data[4:6])
-	r.DataSize = dc.Uint64(data[6:14])
-	r.NameSize = dc.Uint16(data[14:16])
+	r.HeaderSize = int32(dc.Uint32(data[0:4]))
+	r.ReqType = int16(dc.Uint16(data[4:6]))
+	r.DataSize = int64(dc.Uint64(data[6:14]))
+	r.NameSize = int16(dc.Uint16(data[14:16]))
 	size := InitialSize(r.NameSize)
 	if uint64(size) > uint64(n) {
 		return BufferTooSmallError
